【调试】Firmata

+++ component/Firmata/firmata_io.h
@@ -16,7 +16,7 @@ void firmata_io_init();
 void pinMode(uint8_t, uint8_t);
 void digitalWrite(uint8_t, uint8_t);
 int digitalRead(uint8_t);
-float analogRead(uint8_t);
+int analogRead(uint8_t);
 void analogReference(uint8_t mode);
 void analogWrite(uint8_t, int);
